at_the_peak_project/requirements/development.txt
-r base.txt # 
django-extensions
django-rest-swagger